+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Fix issue with inherited hook in helper ([#33](https://github.com/avo-hq/class_variants/pull/33))
+
 ## 1.1.0 (2025-01-20)
 - Add support for merging ([#23](https://github.com/avo-hq/class_variants/pull/23))
 - Add support for subclass inheritance in helper ([#24](https://github.com/avo-hq/class_variants/pull/24))